+++ app/web/wayforge-client/src/services/git.ts
@@ -1,6 +1,6 @@
 import { atom, useRecoilTransaction_UNSTABLE } from "recoil"
 
-import { isGitSocketError } from "~/packages/@git-io/src/git-io"
+import { isGitSocketError } from "~/packages/@git-io/src/core"
 import { initGitClientTools } from "~/packages/@git-io/src/git-io-web"
 import type { Transact } from "~/packages/hamr/recoil-tools/recoil-utils"
