+++ packages/@git-io/src/defaults.ts
@@ -1,4 +1,4 @@
-import type { GitInterface } from "../git-io"
+import type { GitInterface } from "./core"
 
 export const FILLER = `—`
